+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `#[serde(rename_all = "camelCase")]` to enums;
 - Renamed `synced_nodes` to `healthy_nodes`;
 - Renamed `unsynced_nodes()` to `unhealthy_nodes()`;
+- `UrlAuthError(String)` to `UrlAuthError(&'static str)`;
 
 ### Removed